src/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.14
# by the following command:
#
#    pip-compile requirements.in
#
asttokens==3.0.1
    # via stack-data
autopep8==2.3.2
    # via nbqa
black==26.1.0
    # via -r requirements.in
click==8.3.1
    # via black
decorator==5.2.1
    # via ipython
executing==2.2.1
    # via stack-data
ipython==9.10.0
    # via nbqa
ipython-pygments-lexers==1.1.1
    # via ipython
jedi==0.19.2
    # via ipython
matplotlib-inline==0.2.1
    # via ipython
mypy-extensions==1.1.0
    # via black
nbqa==1.9.1
    # via -r requirements.in
packaging==26.0
    # via black
parso==0.8.6
    # via jedi
pathspec==1.0.4
    # via black
pexpect==4.9.0
    # via ipython
platformdirs==4.9.2
    # via black
prompt-toolkit==3.0.52
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pycodestyle==2.14.0
    # via autopep8
pycryptodome==3.23.0
    # via -r requirements.in
pygments==2.19.2
    # via
    #   ipython
    #   ipython-pygments-lexers
pytokens==0.4.1
    # via black
stack-data==0.6.3
    # via ipython
tokenize-rt==6.2.0
    # via nbqa
tomli==2.4.0
    # via nbqa
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
wcwidth==0.6.0
    # via prompt-toolkit
